+++ docs/shared/FieldArrayExampleLoader.tsx
@@ -1,0 +1,5 @@
+import dynamic from 'next/dynamic'
+
+export const FieldArrayExample = dynamic(() => import('./FieldArrayExample'), {
+  ssr: false
+})