+++ docs/shared/SelectWithSubViewLoader.tsx
@@ -1,0 +1,14 @@
+import dynamic from 'next/dynamic'
+
+export const ExampleWidgetSubviewForSelect = dynamic(() => import('./ExampleWidgetSubviewForSelect'), {
+  ssr: false
+})
+
+export const ExampleFormSubviewForSelect = dynamic(() => import('./ExampleFormSubviewForSelect'), {
+  ssr: false
+})
+
+export default {
+  ExampleWidgetSubviewForSelect,
+  ExampleFormSubviewForSelect
+}