+++ docs/shared/StepWizardFormExampleLoader.tsx
@@ -1,0 +1,8 @@
+import dynamic from 'next/dynamic'
+
+export const FormExampleStepWizard = dynamic(
+  () => import('./StepWizardFormExample').then(m => m.FormExampleStepWizard),
+  {
+    ssr: false
+  }
+)