+++ CHANGELOG.md
@@ -2,13 +2,41 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to Semantic Versioning. Use prefixes (Added, Changed, Deprecated, Removed, Fixed, Security) for each change.
 
+## 1.0.165 - 06/17/20
+
+- `Added` tab support (keyboard accessible) for ColorPicker component.
+
+## 1.0.164 - 06/17/20
+
+- `Fixed` LogViewer example .
+
+## 1.0.163 - 06/16/20
+
+- `Fixed` Clicking on Card.Menu also triggers click event on Card itself.
+
+## 1.0.162 - 06/15/20
+
+- `Added` ColorPicker component + its Formik integration.
+
+## 1.0.161 - 06/12/20
+
+- `Changed` Collapse header props are sent to collapse list panel.
+- `Changed` Children for collapse list header now has a type.
+
+## 1.0.160 - 06/12/20
+
+- `Fixed` undefined window when integrating third party libs.
+- `Changed` Masonry props to expose ref.
+
 ## 1.0.159 - 06/12/20
 
 - `Fixed` Set Auto Complete default to OFF for formik text and text area field.
+- `Added` `<LogViewer />` component to view logs.
 
 ## 1.0.158 - 06/12/20
 
-- `Added` `<LogViewer />` component to view logs.
+- `Added` Masonry's prop and type to UIKit exports.
+- `Fixed` `<Layout.Masonry/>` not updating layout when data got changed.
 
 ## 1.0.157 - 06/12/20